# Simplify ElementListTagNameMap