--- a/TS.fsx
+++ b/TS.fsx
@@ -824,10 +824,11 @@ module Emit =
         Pt.Printl ""
 
     let EmitElementTagNameMap () =
-        Pt.Printl "interface ElementTagNameMap {"
+        Pt.Printl "interface ElementTagNameMap extends HTMLElementTagNameMap {"
         Pt.IncreaseIndent()
         for e in tagNameToEleName do
-            Pt.Printl "\"%s\": %s;" (e.Key.ToLower()) e.Value
+            if iNameToIDependList.ContainsKey e.Value && not (Seq.contains "HTMLElement" iNameToIDependList.[e.Value]) then
+                Pt.Printl "\"%s\": %s;" (e.Key.ToLower()) e.Value
         Pt.DecreaseIndent()
         Pt.Printl "}"
         Pt.Printl ""