This replaces long map with a simple indexer. No performance tradeoff.

Original:

```
Files:            2
Lines:        19215
Nodes:        79623
Identifiers:  27017
Symbols:      20532
Types:         5505
Memory used: 51801K
I/O read:     0.00s
I/O write:    0.00s
Parse time:   0.33s
Bind time:    0.14s
Check time:   0.59s
Emit time:    0.02s
Total time:   1.07s
```

This PR:

```
Files:            2
Lines:        18954
Nodes:        78233
Identifiers:  26582
Symbols:      20272
Types:         5386
Memory used: 51598K
I/O read:     0.00s
I/O write:    0.00s
Parse time:   0.32s
Bind time:    0.13s
Check time:   0.59s
Emit time:    0.03s
Total time:   1.07s
```